Moved component from module to theme directory
