--- src/modules/vsf-back-to-top/index.ts
@@ -1,6 +1,0 @@
-import { createModule } from '@vue-storefront/core/lib/module'
-
-const KEY = 'vsf-back-to-top'
-export const BackToTop = createModule({
-  key: KEY
-})
